added new channels, improved existing file check

+++ main.py
@@ -10,7 +10,7 @@ class CollectChannelData:
     def update(self):
         channel = Channel(self.channel_url)
         channel_name = channel.channel_name
-        filename= "../DATA/"+ channel_name + "_manuel.csv"
+        filename= "../DATA/"+ channel_name + "_manuell.csv"
         try:
             print(filename)
             df = pd.read_csv(filename)
@@ -24,43 +24,49 @@ class CollectChannelData:
 
 
     def create_empty_csv(self, features):
-        #TODO: Excetion handeln.... csv File überschreiben.....
- 
         channel = Channel(self.channel_url)
         channel_name = channel.channel_name
         basefolder = r"\\DATEN\Schöpferwissen\.DATA"
         filename = os.path.join(basefolder, channel_name + "_manuell.csv")
-        try:
-            df = pd.DataFrame(filename)
-            print("File already exists. Use self.update for adding")
-        except:
-            df = pd.DataFrame(columns=features)
-            listofblanks = [' '] * len(features)
-            for url in channel.video_urls:
-                video_id = url.replace("https://www.youtube.com/watch?v=", "")
-                print(video_id)
-                new_row = listofblanks
-                new_row[0] = video_id
-                df.loc[len(df.index)]= new_row
-            df.to_csv(filename, index=False)
-            print("NEW Filename: " + filename)
+
+        if os.path.exists(filename):
+            print(f'File {filename} already exists. Skipping.')
+            return
+
+        df = pd.DataFrame(columns=features)
+        listofblanks = [' '] * len(features)
+        print(f"Processing manual CSV for channel {channel_name}...")
+        for url in channel.video_urls:
+            video_id = url.replace("https://www.youtube.com/watch?v=", "")
+            print(video_id, "\r")
+            new_row = listofblanks
+            new_row[0] = video_id
+            df.loc[len(df.index)]= new_row
+        df.to_csv(filename, index=False)
+        print("="*100)
+        print("Saved as:", filename)
 
 
     def process_channel(self):
-        #TODO: Excetion handeln.... csv File überschreiben.....
-
         df = pd.DataFrame()
         channel = Channel(self.channel_url)
         channel_name = channel.channel_name
-        filename= "DATA/"+ channel_name + "_automated.csv"
+        basefolder = r"\\DATEN\Schöpferwissen\.DATA"
+        filename = os.path.join(basefolder, channel_name + "_automated.csv")
+
+        if os.path.exists(filename):
+            print(f'File {filename} already exists. Skipping.')
+            return
+
         features = ["publish_date", "id", "title", "description", "keywords", "length", "views", "age_restricted", "yt_caption_info", "yt_caption_tracks", "vid_info"]
         df = pd.DataFrame(columns=features)
         new_row = [0] * len(features)
+        print(f"Processing channel {channel_name} for metadata...")
         for url in channel.video_urls:
             yt = YouTube(url)
             yt.bypass_age_gate()
             yt.check_availability()
-            
+            print(video_id, "\r")
             publish_date = yt.publish_date
             id = url.replace("https://www.youtube.com/watch?v=", "")
             title = yt.title
@@ -85,7 +91,7 @@ class CollectChannelData:
             new_row[9] = vid_info
             df.loc[len(df.index)]= new_row
         print("="*100)
-        print("Saved as:" + filename)
+        print("Saved as:", filename)
         print("="*100)
         df.to_csv(filename, index=False)
             
@@ -109,6 +115,13 @@ if __name__ == '__main__':
     Drachentoeter_channel_url="https://www.youtube.com/channel/UC4StsLMnfcOuQ374_oGa5sg/videos"
     #Nebenkanal "UBC TV": 11 Videos
     UBC_channel_url="https://www.youtube.com/channel/UCWELnCGV_IYWctpTPiB8-Sw"
+    #Nebenkanal "u n endlich"
+    unendlich_channel_url="https://www.youtube.com/channel/UCyIj6T2W5xQsi7dCbfuJdUw"
+    #Nebenkanal "Honett"
+    Honett_channel_url="https://www.youtube.com/channel/UC2BK1JGDSB0RFSJdijEBnpA"
+    #Nebenkanal "WISSEN - NICHT VON DIESER WELT"
+    WISSEN_channel_url="https://www.youtube.com/channel/UC5Sr_xoXTkBThtqthLwOkPg"
+
     all_channels = [
         Schoepf_channel_url,
         Wahrheit_channel_url,
@@ -117,7 +130,10 @@ if __name__ == '__main__':
         ALLES_channel_url,
         VEITCLUB_channel_url,
         Drachentoeter_channel_url,
-        UBC_channel_url
+        UBC_channel_url,
+        unendlich_channel_url,
+        Honett_channel_url,
+        WISSEN_channel_url
     ]
 
     #EINE CSV ERSTELLEN
@@ -129,7 +145,7 @@ if __name__ == '__main__':
         try:
             collect_channel_data = CollectChannelData(item)
             collect_channel_data.process_channel()
-            #collect_channel_data.create_empty_csv(features)
+            collect_channel_data.create_empty_csv(features)
         except:
             print(item +" hat nicht geklappt")
     print("FERTIG HAT ALLES GEKLAPPT")
